refactor(templates): move registry cache from module singleton to Pinia store
useTemplateResolver held template_variables in a module-level `registryCache`
`let` with a test-only `__resetTemplateRegistryCache` export — a private-impl
leak that forced tests to reach into module state.

Move the cache into stores/templateRegistry.ts (Pinia). loadRegistry now
delegates to the store's `load(force)`; the cache is devtools-visible and resets
per test via a fresh Pinia instead of an exported reset hook. Same fetch,
PGRST205 graceful-degrade, and fail-open behavior.

Spec swaps `__resetTemplateRegistryCache()` for `setActivePinia(createPinia())`;
the cache-once assertion (second load doesn't refetch) still holds.

type-check + lint clean; 7976 unit tests pass.
